// packages/js-core/src/proto/polycentric.ts
export interface PublicKey {
  /**
   * @generated from protobuf field: uint64 key_type = 1
   */
  keyType: bigint;
  /**
   * @generated from protobuf field: bytes key = 2
   */
  key: Uint8Array;
}
// ...
export interface Index {
  /**
   * @generated from protobuf field: uint64 index_type = 1
   */
  indexType: bigint;
  /**
   * @generated from protobuf field: uint64 logical_clock = 2
   */
  logicalClock: bigint;
}
/**
 * @generated from protobuf message polycentric.Indices
 */
export interface Indices {
  /**
   * @generated from protobuf field: repeated polycentric.Index indices = 1
   */
  indices: Index[];
}
/**
 * @generated from protobuf message polycentric.Process
 */
export interface Process {
  /**
   * @generated from protobuf field: bytes process = 1
   */
  process: Uint8Array;
}
/**
 * @generated from protobuf message polycentric.Digest
 */
export interface Digest {
  /**
   * @generated from protobuf field: uint64 digest_type = 1
   */
  digestType: bigint;
  /**
   * @generated from protobuf field: bytes digest = 2
   */
  digest: Uint8Array;
}
/**
 * @generated from protobuf enum polycentric.Digest.DigestType
 */
export enum Digest_DigestType {
  /**
   * @generated from protobuf enum value: Unknown = 0;
   */
  Unknown = 0,
  /**
   * @generated from protobuf enum value: SHA256 = 1;
   */
  SHA256 = 1,
}
/**
 * @generated from protobuf message polycentric.Pointer
 */
export interface Pointer {
  /**
   * @generated from protobuf field: polycentric.PublicKey system = 1
   */
  system?: PublicKey;
  /**
   * @generated from protobuf field: polycentric.Process process = 2
   */
  process?: Process;
  /**
   * @generated from protobuf field: uint64 logical_clock = 3
   */
  logicalClock: bigint;
  /**
   * @generated from protobuf field: polycentric.Digest event_digest = 4
   */
  eventDigest?: Digest;
}
/**
 * @generated from protobuf message polycentric.LWWElementSet
 */
export interface LWWElementSet {
  /**
   * @generated from protobuf field: polycentric.LWWElementSet.Operation operation = 1
   */
  operation: LWWElementSet_Operation;
  /**
   * @generated from protobuf field: bytes value = 2
   */
  value: Uint8Array;
  /**
   * @generated from protobuf field: uint64 unix_milliseconds = 3
   */
  unixMilliseconds: bigint;
}
/**
 * @generated from protobuf enum polycentric.LWWElementSet.Operation
 */
export enum LWWElementSet_Operation {
  /**
   * @generated from protobuf enum value: ADD = 0;
   */
  ADD = 0,
  /**
   * @generated from protobuf enum value: REMOVE = 1;
   */
  REMOVE = 1,
}
/**
 * @generated from protobuf message polycentric.LWWElement
 */
export interface LWWElement {
  /**
   * @generated from protobuf field: bytes value = 1
   */
  value: Uint8Array;
  /**
   * @generated from protobuf field: uint64 unix_milliseconds = 2
   */
  unixMilliseconds: bigint;
}
/**
 * @generated from protobuf message polycentric.Reference
 */
export interface Reference {
  /**
   * @generated from protobuf field: uint64 reference_type = 1
   */
  referenceType: bigint;
  /**
   * @generated from protobuf field: bytes reference = 2
   */
  reference: Uint8Array;
}
// ...
export interface ModerationTag {
  /**
   * @generated from protobuf field: string name = 1
   */
  name: string;
  /**
   * @generated from protobuf field: uint32 level = 2
   */
  level: number;
}
/**
 * @generated from protobuf message polycentric.SignedEvent
 */
export interface SignedEvent {
  /**
   * @generated from protobuf field: bytes signature = 1
   */
  signature: Uint8Array;
  /**
   * @generated from protobuf field: bytes event = 2
   */
  event: Uint8Array;
  /**
   * @generated from protobuf field: repeated polycentric.ModerationTag moderation_tags = 3
   */
  moderationTags: ModerationTag[];
}
// ...
export interface Claim {
  /**
   * @generated from protobuf field: uint64 claim_type = 1
   */
  claimType: bigint;
  /**
   * @generated from protobuf field: repeated polycentric.ClaimFieldEntry fields = 2
   */
  fields: ClaimFieldEntry[];
}
/**
 * @generated from protobuf message polycentric.ClaimFieldEntry
 */
export interface ClaimFieldEntry {
  /**
   * @generated from protobuf field: uint64 key = 1
   */
  key: bigint;
  /**
   * @generated from protobuf field: string value = 2
   */
  value: string;
}
// ...
export interface EventCreationData {
  /**
   * @generated from protobuf field: polycentric.PublicKey system = 1
   */
  system?: PublicKey;
  /**
   * @generated from protobuf field: polycentric.Process process = 2
   */
  process?: Process;
  /**
   * @generated from protobuf field: polycentric.ContentType content_type = 3
   */
  contentType: ContentType;
  /**
   * @generated from protobuf field: bytes content = 4
   */
  content: Uint8Array;
  /**
   * @generated from protobuf field: optional polycentric.Indices indices = 5
   */
  indices?: Indices;
  /**
   * @generated from protobuf field: optional polycentric.LWWElementSet lww_element_set = 6
   */
  lwwElementSet?: LWWElementSet;
  /**
   * @generated from protobuf field: optional polycentric.LWWElement lww_element = 7
   */
  lwwElement?: LWWElement;
  /**
   * @generated from protobuf field: repeated polycentric.Reference references = 8
   */
  references: Reference[];
  /**
   * @generated from protobuf field: optional uint64 unix_milliseconds = 9
   */
  unixMilliseconds?: bigint;
  /**
   * @generated from protobuf field: optional uint64 logical_clock = 10
   */
  logicalClock?: bigint;
}
// ...
export enum ContentType {
  /**
   * @generated from protobuf enum value: UNKNOWN = 0;
   */
  UNKNOWN = 0,
  /**
   * @generated from protobuf enum value: DELETE = 1;
   */
  DELETE = 1,
  /**
   * @generated from protobuf enum value: SYSTEM_PROCESSES = 2;
   */
  SYSTEM_PROCESSES = 2,
  /**
   * @generated from protobuf enum value: POST = 3;
   */
  POST = 3,
  /**
   * @generated from protobuf enum value: FOLLOW = 4;
   */
  FOLLOW = 4,
  /**
   * @generated from protobuf enum value: USERNAME = 5;
   */
  USERNAME = 5,
  /**
   * @generated from protobuf enum value: DESCRIPTION = 6;
   */
  DESCRIPTION = 6,
  /**
   * @generated from protobuf enum value: BLOB_META = 7;
   */
  BLOB_META = 7,
  /**
   * @generated from protobuf enum value: BLOB_SECTION = 8;
   */
  BLOB_SECTION = 8,
  /**
   * @generated from protobuf enum value: AVATAR = 9;
   */
  AVATAR = 9,
  /**
   * @generated from protobuf enum value: SERVER = 10;
   */
  SERVER = 10,
  /**
   * @generated from protobuf enum value: VOUCH = 11;
   */
  VOUCH = 11,
  /**
   * @generated from protobuf enum value: CLAIM = 12;
   */
  CLAIM = 12,
  /**
   * @generated from protobuf enum value: BANNER = 13;
   */
  BANNER = 13,
  /**
   * @generated from protobuf enum value: OPINION = 14;
   */
  OPINION = 14,
  /**
   * @generated from protobuf enum value: STORE = 15;
   */
  STORE = 15,
  /**
   * @generated from protobuf enum value: AUTHORITY = 16;
   */
  AUTHORITY = 16,
  /**
   * @generated from protobuf enum value: JOIN_TOPIC = 17;
   */
  JOIN_TOPIC = 17,
  /**
   * @generated from protobuf enum value: BLOCK = 18;
   */
  BLOCK = 18,
}
// ...
export const PublicKey = new PublicKey$Type();
// ...
export const Index = new Index$Type();
// ...
export const Indices = new Indices$Type();
// ...
export const Process = new Process$Type();
// ...
export const Digest = new Digest$Type();
// ...
export const Pointer = new Pointer$Type();
// ...
export const LWWElementSet = new LWWElementSet$Type();
// ...
export const LWWElement = new LWWElement$Type();
// ...
export const Reference = new Reference$Type();
// ...
export const ModerationTag = new ModerationTag$Type();
// @generated message type with reflection information, may provide speed optimized methods
class SignedEvent$Type extends MessageType<SignedEvent> {
  constructor() {
    super('polycentric.SignedEvent', [
      { no: 1, name: 'signature', kind: 'scalar', T: 12 /*ScalarType.BYTES*/ },
      { no: 2, name: 'event', kind: 'scalar', T: 12 /*ScalarType.BYTES*/ },
      {
        no: 3,
        name: 'moderation_tags',
        kind: 'message',
        repeat: 2 /*RepeatType.UNPACKED*/,
        T: () => ModerationTag,
      },
    ]);
  }
  create(value?: PartialMessage<SignedEvent>): SignedEvent {
    const message = globalThis.Object.create(this.messagePrototype!);
    message.signature = new Uint8Array(0);
    message.event = new Uint8Array(0);
    message.moderationTags = [];
    if (value !== undefined)
      reflectionMergePartial<SignedEvent>(this, message, value);
    return message;
  }
  internalBinaryRead(
    reader: IBinaryReader,
    length: number,
    options: BinaryReadOptions,
    target?: SignedEvent,
  ): SignedEvent {
    let message = target ?? this.create(),
      end = reader.pos + length;
    while (reader.pos < end) {
      let [fieldNo, wireType] = reader.tag();
      switch (fieldNo) {
        case /* bytes signature */ 1:
          message.signature = reader.bytes();
          break;
        case /* bytes event */ 2:
          message.event = reader.bytes();
          break;
        case /* repeated polycentric.ModerationTag moderation_tags */ 3:
          message.moderationTags.push(
            ModerationTag.internalBinaryRead(reader, reader.uint32(), options),
          );
          break;
        default:
          let u = options.readUnknownField;
          if (u === 'throw')
            throw new globalThis.Error(
              `Unknown field ${fieldNo} (wire type ${wireType}) for ${this.typeName}`,
            );
          let d = reader.skip(wireType);
          if (u !== false)
            (u === true ? UnknownFieldHandler.onRead : u)(
              this.typeName,
              message,
              fieldNo,
              wireType,
              d,
            );
      }
    }
    return message;
  }
  internalBinaryWrite(
    message: SignedEvent,
    writer: IBinaryWriter,
    options: BinaryWriteOptions,
  ): IBinaryWriter {
    /* bytes signature = 1; */
    if (message.signature.length)
      writer.tag(1, WireType.LengthDelimited).bytes(message.signature);
    /* bytes event = 2; */
    if (message.event.length)
      writer.tag(2, WireType.LengthDelimited).bytes(message.event);
    /* repeated polycentric.ModerationTag moderation_tags = 3; */
    for (let i = 0; i < message.moderationTags.length; i++)
      ModerationTag.internalBinaryWrite(
        message.moderationTags[i],
        writer.tag(3, WireType.LengthDelimited).fork(),
        options,
      ).join();
    let u = options.writeUnknownFields;
    if (u !== false)
      (u == true ? UnknownFieldHandler.onWrite : u)(
        this.typeName,
        message,
        writer,
      );
    return writer;
  }
}
/**
 * @generated MessageType for protobuf message polycentric.SignedEvent
 */
export const SignedEvent = new SignedEvent$Type();
// ...
export const Claim = new Claim$Type();
// @generated message type with reflection information, may provide speed optimized methods
class ClaimFieldEntry$Type extends MessageType<ClaimFieldEntry> {
  constructor() {
    super('polycentric.ClaimFieldEntry', [
      {
        no: 1,
        name: 'key',
        kind: 'scalar',
        T: 4 /*ScalarType.UINT64*/,
        L: 0 /*LongType.BIGINT*/,
      },
      { no: 2, name: 'value', kind: 'scalar', T: 9 /*ScalarType.STRING*/ },
    ]);
  }
  create(value?: PartialMessage<ClaimFieldEntry>): ClaimFieldEntry {
    const message = globalThis.Object.create(this.messagePrototype!);
    message.key = 0n;
    message.value = '';
    if (value !== undefined)
      reflectionMergePartial<ClaimFieldEntry>(this, message, value);
    return message;
  }
  internalBinaryRead(
    reader: IBinaryReader,
    length: number,
    options: BinaryReadOptions,
    target?: ClaimFieldEntry,
  ): ClaimFieldEntry {
    let message = target ?? this.create(),
      end = reader.pos + length;
    while (reader.pos < end) {
      let [fieldNo, wireType] = reader.tag();
      switch (fieldNo) {
        case /* uint64 key */ 1:
          message.key = reader.uint64().toBigInt();
          break;
        case /* string value */ 2:
          message.value = reader.string();
          break;
        default:
          let u = options.readUnknownField;
          if (u === 'throw')
            throw new globalThis.Error(
              `Unknown field ${fieldNo} (wire type ${wireType}) for ${this.typeName}`,
            );
          let d = reader.skip(wireType);
          if (u !== false)
            (u === true ? UnknownFieldHandler.onRead : u)(
              this.typeName,
              message,
              fieldNo,
              wireType,
              d,
            );
      }
    }
    return message;
  }
  internalBinaryWrite(
    message: ClaimFieldEntry,
    writer: IBinaryWriter,
    options: BinaryWriteOptions,
  ): IBinaryWriter {
    /* uint64 key = 1; */
    if (message.key !== 0n) writer.tag(1, WireType.Varint).uint64(message.key);
    /* string value = 2; */
    if (message.value !== '')
      writer.tag(2, WireType.LengthDelimited).string(message.value);
    let u = options.writeUnknownFields;
    if (u !== false)
      (u == true ? UnknownFieldHandler.onWrite : u)(
        this.typeName,
        message,
        writer,
      );
    return writer;
  }
}
/**
 * @generated MessageType for protobuf message polycentric.ClaimFieldEntry
 */
export const ClaimFieldEntry = new ClaimFieldEntry$Type();
// ...
export const EventCreationData = new EventCreationData$Type();